requirements.txt
Flask==2.3.3
Flask-Mail==0.9.1
Flask-WTF==1.1.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
redis==5.0.1
scikit-learn==1.4.1
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
email-validator==2.1.0.post1
Flask-Session==0.4.0
